Updated requred cmake version to 3.1
FIxed possible compilation error - added inclusion of <stdexcept> to
ClassField.h.
Fixed couple of remained errors in connection and bugs tests, where properties
were initialized with bool value.
Also, fixed few compilation warnings(CONCPP-12). 2 of them weren't
harmless and were bugs(assingment instead of comparison in 'if'
condition).

+++ src/ColumnDefinition.cpp
@@ -43,8 +43,8 @@ namespace mariadb
 
     md.name= (char*)name.c_str();
     md.org_name= (char*)name.c_str();
-    md.name_length= name.length();
-    md.org_name_length= name.length();
+    md.name_length= static_cast<unsigned int>(name.length());
+    md.org_name_length= static_cast<unsigned int>(name.length());
 
     switch (_type.getSqlType()) {
     case Types::VARCHAR:
